pyproject.toml
[tool.poetry]
name = "inkal"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.149.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
pytz = "^2024.2"
selenium = "^4.25.0"
pillow = "^10.4.0"
rpi-gpio = "^0.7.1"
spidev = "^3.6"
typings-google-calendar-api = "^1.0.0"
pytest = "^8.3.3"
htmlgenerator = "^1.2.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
